Read a table as data frame

Convert a table to a temporary text file pass to readr::read_delim().

read_mdb(
  file,
  table = NULL,
  delim = ",",
  quote = "\"",
  quote_escape = "\"",
  col_names = TRUE,
  date_format = "%Y-%m-%d %H:%M:%S",
  col_types = mdb_schema(file, table),
  ...
)

Arguments

file
Path to the Microsoft Access file.

table
Name of the table, use mdb_tables().

delim
Single character used to separate fields within a record.

quote
Single character used to quote strings. Defaults to ".

quote_escape
A single character (or empty string) to escape double quotes with text columns. Defaults to doubling although backslashes are also used.

col_names
Whether or not to suppress column names from the table.

date_format
The format in which date columns are converted. MDB Tools uses the strftime(3) format, similar to readr::parse_date(). No need to specify whole string. Defaults to ISO8601.

col_types
One of NULL, a readr::cols() specification, or a string. See vignette("readr") for more details. A generic readr::cols() specification can be made by mdb_schema() via the mdb-schema utility.

...
Additional arguments passed to readr::read_delim().

Value

A data frame.

Examples

read_mdb(mdb_example(), "Flights")

#> Warning: error in running command

#> Warning: `data_frame()` is deprecated as of tibble 1.1.0.
#> Please use `tibble()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> # A tibble: 0 x 0